+++ docs/plugins/ui-components.ts
@@ -1,6 +1,0 @@
-import { UButton } from '#components'
-
-// Expose Nuxt UI components to Nuxt Content
-export default defineNuxtPlugin((nuxtApp) => {
-  nuxtApp.vueApp.component('UButton', UButton)
-})